src/evaluation.py
class Evaluator:
    """Evaluates the performance of question generation models.

    This class provides methods for calculating metrics like BLEU score, ROUGE score, etc.
    """

    def __init__(self):
        """Initializes the Evaluator.

        Currently, no specific initialization is required.
        """
        pass

    def calculate_bleu(self, candidate, reference):
        """Calculates the BLEU score between a candidate and a reference sentence.

        Args:
            candidate (str): The generated question (candidate).
            reference (str): The ground truth question (reference).

        Returns:
            float: The BLEU score.
        """
        # Implement BLEU score calculation here
        # This is a placeholder; replace with actual BLEU calculation
        return 0.0

    def calculate_rouge(self, candidate, reference):
       """Calculates the ROUGE score between a candidate and a reference sentence.

       Args:
           candidate (str): The generated question (candidate).
           reference (str): The ground truth question (reference).

       Returns:
           float: The ROUGE score.
       """
       # Implement ROUGE score calculation here
       # This is a placeholder; replace with actual ROUGE calculation
       return 0.0